add transport inbound, create raport, fix bugs

+++ src/components/EditForm/EditForm.js
@@ -2,7 +2,6 @@ import React, { useEffect, useState } from "react";
 import Box from "../Box/Box";
 import Form from "../Form/Form";
 import Title from "../Common/Title/Title";
-import Input from "../Input/Input";
 import Button from "../Button/Button";
 import { useNavigate, useParams } from "react-router-dom";
 
@@ -21,19 +20,36 @@ import {
 import db from "../../services/firebase/firebaseConfig";
 import { deleteDoc, doc, getDoc, updateDoc } from "firebase/firestore";
 import Notification from "../Common/Notification/Notification";
+import TextInput from "../Common/TextInput/TextInput";
 
 
-const EditForm = () => {
+const EditForm = ({type}) => {
   const { id } = useParams();
   const navigation = useNavigate();
   const [showModal, setShowModal] = useState(false);
   const [notification, setNotification] = useState({ message: "", type: "" });
-
   const [rowData, setRowData] = useState({});
+  let navigatePath;
+  let collectionName;
+
+  if(type === 'outbound'){
+    collectionName = 'transport';
+    navigatePath = '/listTransportOutbound';
+  }else if(type === 'inbound'){
+    collectionName = 'TransportInbound';
+    navigatePath = '/listTransportInbound';
+  }else{
+    console.log('invalid type');
+  }
 
   useEffect(() => {
     const fetchData = async () => {
-      const docRef = await getDoc(doc(db, "transport", id));
+      if (!collectionName || !id) {
+        console.log('collectionName or id is undefined');
+        return;
+      }
+  
+      const docRef = await getDoc(doc(db, collectionName, id));
       if (docRef.exists()) {
         setRowData(docRef.data());
       } else {
@@ -41,11 +57,30 @@ const EditForm = () => {
       }
     };
     fetchData();
-  }, [id]);
+  }, [id, type]);
 
-  const handlePrint = () => {
-    const doc = generatePdf(rowData);
-    doc.output("dataurlnewwindow"); // Open PDF in a new window
+  const handlePrint = async() => {
+    // Pobierz dane przewoźnika z bazy danych
+  const carrierRef = doc(db, 'Carriers', rowData.Carrier);
+  const carrierSnap = await getDoc(carrierRef);
+
+  if (!carrierSnap.exists()) {
+    console.log('No such carrier!');
+    return;
+  }
+
+    const pdfDoc = generatePdf(rowData);
+    pdfDoc.output("dataurlnewwindow"); // Open PDF in a new window
+    try {
+      const docRef = doc(db, collectionName, id);
+      await updateDoc(docRef, {
+        Status: "Closed"
+      });
+      setRowData(prevRow => ({...prevRow, Status: 'Closed'}));
+    } catch (e) {
+      console.log(e.message);
+    }
+    navigation(navigatePath);
   };
 
   const handleChange = (e) => {
@@ -56,9 +91,15 @@ const EditForm = () => {
   };
 
   const handleSave = async () => {
+    if (!collectionName || !id) {
+      console.log('collectionName or id is undefined');
+      return;
+    }
+  
     try {
-      const docRef = doc(db, "transport", id);
+      const docRef = doc(db, collectionName, id);
       await updateDoc(docRef, {
+       Status: rowData.Status,
         Carrier: rowData.Carrier,
         Carrier_Number: rowData.Carrier_Number,
         License_Truck: rowData.License_Truck,
@@ -78,9 +119,14 @@ const EditForm = () => {
   };
 
   const handleDelete = async (id) => {
-    const docRef = doc(db, "transport", id);
-    await deleteDoc(docRef);
-    navigation(`/listTransport`);
+    if (!collectionName || !id) {
+      console.log('collectionName or id is undefined');
+      return;
+    }
+
+      const docRef = doc(db, collectionName, id);
+      await deleteDoc(docRef);
+      navigation(navigatePath);
   };
 
   return (
@@ -94,7 +140,7 @@ const EditForm = () => {
             <div className="w-full mt-8 mx-auto px-16 py-8 rounded-lg">
               <div className="flex justify-between items-center">
               <Title tag="h3">Edit transport</Title>
-              <Button onClick={() => navigation(`/listTransport`)}>Back</Button>
+              <Button onClick={() => navigation(navigatePath)}>Back</Button>
               </div>
               <Form
                 onSubmit={(e) => {
@@ -106,7 +152,7 @@ const EditForm = () => {
                   <div className="flex flex-wrap -mx-3">
                     <div className="w-full md:w-1/2  mb-6 md:mb-0">
                       <div className="my-4 mx-2">
-                        <Input
+                        <TextInput
                           type="text"
                           icon={TruckIcon}
                           name="Carrier"
@@ -116,7 +162,7 @@ const EditForm = () => {
                         />
                       </div>
                       <div className="my-4 mx-2">
-                        <Input
+                        <TextInput
                           type="text"
                           icon={TruckIcon}
                           name="Carrier_Number"
@@ -126,7 +172,7 @@ const EditForm = () => {
                         />
                       </div>
                       <div className="my-4 mx-2">
-                        <Input
+                        <TextInput
                           type="text"
                           icon={PagerIcon}
                           name="Pager"
@@ -138,7 +184,7 @@ const EditForm = () => {
                     </div>
                     <div className="w-full md:w-1/2">
                       <div className="my-4 mx-2">
-                        <Input
+                        <TextInput
                           type="text"
                           icon={TruckFrontIcon}
                           name="License_Truck"
@@ -148,8 +194,7 @@ const EditForm = () => {
                         />
                       </div>
                       <div className="my-4 mx-2">
-                        <Input
-                          type="text"
+                        <TextInput
                           icon={TrailerIcon}
                           name="License_Trailer"
                           label="License trailer"
@@ -163,8 +208,7 @@ const EditForm = () => {
                   <div className="flex flex-wrap -mx-3">
                     <div className="w-full md:w-1/2  mb-6 md:mb-0">
                       <div className="my-4 mx-2">
-                        <Input
-                          type="text"
+                        <TextInput
                           icon={PalletIcon}
                           name="Pallets"
                           label="Number of pallets"
@@ -173,8 +217,7 @@ const EditForm = () => {
                         />
                       </div>
                       <div className="my-4 mx-2">
-                        <Input
-                          type="text"
+                        <TextInput
                           icon={TruckIcon}
                           name="Seal"
                           label="Seal"
@@ -183,19 +226,18 @@ const EditForm = () => {
                         />
                       </div>
                       <div className="my-4 mx-2">
-                        <Input
-                          type="text"
+                        <TextInput
                           icon={CalendarIcon}
                           name="Departure"
                           label="Departure date"
                           value={rowData?.Departure}
+                          onChange={handleChange}
                         />
                       </div>
                     </div>
                     <div className="w-full md:w-1/2 ">
                       <div className="my-4 mx-2">
-                        <Input
-                          type="text"
+                        <TextInput
                           icon={PackageIcon}
                           name="Package"
                           label="Number of packages"
@@ -204,8 +246,7 @@ const EditForm = () => {
                         />
                       </div>
                       <div className="my-4 mx-2">
-                        <Input
-                          type="text"
+                        <TextInput
                           icon={WeightIcon}
                           name="Weight"
                           label="Weight"
@@ -215,7 +256,7 @@ const EditForm = () => {
                       </div>
                     </div>
                     <div className="w-full mx-2">
-                      <Input
+                      <TextInput
                         type="textarea"
                         tag="textarea"
                         name="Message"
@@ -229,7 +270,6 @@ const EditForm = () => {
 
                 <div className="flex w-full">
                   <Button
-                    
                     type="submit"
                     className="grow mx-3 mt-4"
                     onClick={handleSave}
@@ -239,7 +279,8 @@ const EditForm = () => {
                   <Button
                     bg="bg-sky-700"
                     onClick={() => handlePrint()}
-                    className="grow mx-3 mt-4"
+                    className={`grow mx-3 mt-4 ${!rowData.Carrier || !rowData.Carrier_Number || !rowData.License_Truck || !rowData.License_Trailer || !rowData.Pager || !rowData.Pallets || !rowData.Seal || !rowData.Departure || !rowData.Package || !rowData.Weight ? 'bg-blue-gray-800 hover:bg-blue-gray-800 cursor-not-allowed' : ''}`}
+                    disabled={!rowData.Carrier || !rowData.Carrier_Number || !rowData.License_Truck || !rowData.License_Trailer || !rowData.Pager || !rowData.Pallets || !rowData.Seal || !rowData.Departure || !rowData.Package || !rowData.Weight}
                   >
                     Print
                   </Button>
